\name{AQLSchemes-package}
\alias{AQLSchemes-package}
\docType{package}
\title{
This package contains functions to recall MIL-STD-105E-ANSI/ASQ Z1.4, and the MIL-STD-414-ANSI/ASQ Z1.9  
Variable Sampling Plans from the Public Domain Published Tables 
}
\description{The \pkg{AAQLSchemes} package provides a functions to recall MIL-STD-105E-ANSI/ASQ Z1.4, MIL-STD-414-ANSI Z1.9
}

\details{
\tabular{ll}{
Package: \tab AQLSchemes\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2018-07-06\cr
License: \tab GPL2.0\cr
Dependencies: \tab \cr
LazyLoad: \tab yes\cr
Packaged: \tab 2018-07-06 19:54:07 UTC; Lawson\cr
Built: \tab R 3.5.1; i386-pc-mingw32; 2011-03-22 19:54:08 UTC; windows\cr
}

Index:
\preformatted{
AADouble                Calls function AAZ14Double
AAMultiple              Calls function AAZ14Multiple
AASingle                Calls function AAZ14Single
AAZ14Double             Recalls and prints Sampling plans
AAZ14Multiple           Recalls and prints Sampling Plans
AAZ14Single             Recalls and prints Sampling Plans
AAZ19                   Recalls and Prints Z19 Sampling Plans
EPn                     Calulates the estimated proprtion non-conforming
}
}
\author{
John Lawson <lawson@byu.edu>

Maintainer: John Lawson <lawson@byu.edu>
}
\keyword{ package }
